% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCGImpute.R
\name{scCGImpute}
\alias{scCGImpute}
\title{Title}
\usage{
scCGImpute(count_matrix, out_dir, Kcluster = NULL)
}
\arguments{
\item{count_matrix}{A matrix specifying the scRNA-seq data}

\item{out_dir}{A character specifying the path of the output directory}

\item{Kcluster}{An integer specifying the number of cell subpopulations}
}
\description{
Title
}
